image_recognition-mnist-cnn-classifier/requirements.txt
# requirements.txt
Flask
flasgger
numpy
Pillow
tensorflow-cpu
opencv-python
Werkzeug
matplotlib
scipy
